[AOTI] Extend torchgen to generate C shim with version number
Summary: While it is ok to add a new arg with defaul value to a fallback op in Python, it will be BC-breaking for the C shim. This PR adds an automatic approach to update C shim files when specifying a version number with a list of new args for the modified op. See https://github.com/pytorch/pytorch/pull/154848 as an example on how to do that.

ghstack-source-id: c67964fe5024669c33dd1c441a9bb5309872f3c6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/147745