[inductor][test] test_padding.py: use inductor TestCase instead of dynamo TestCase
test_pad_3d_tensor fails if you run it multiple times in a row, because the cache is populated and inductor skips the logic that increments the counter.

To fix this, switch these tests to use inductor's TestCase / run_tests instead of dynamo's - this way, a fresh inductor cache is used.

ghstack-source-id: b29d499c8f6040249d1c79ba99e2f5fe062f7ea2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/154935