Improve safetensors consolidation script
This improves the safetensor consolidation script by adding the ability to multi thread so that we can write to multiple files concurrently and speed up the consolidation process. It also can take in a safetensors metadata file for the mapping of fqns to what files they should be part of which will help with parallelization. It also changes the full deserialization, to only read tensor by tensor. Finally, there is an upload method so that users can upload the output files to remote storage from the script

Differential Revision: [D75790351](https://our.internmc.facebook.com/intern/diff/D75790351/)

ghstack-source-id: 287652218
Pull Request resolved: https://github.com/pytorch/pytorch/pull/155706