[dtensor] refactor PlacementStrategy -> OpSpec, move utils to OpSchema
as titled. It's sometimes confusing to use PlacementStrategy as a name,
as we also have OpStrategy and TupleStrategy, the latter two contain
the former, so it is better to make the naming clearer.

Renaming PlacementStrategy -> OpSpec as it is an operator spec that
contains output_spec + input_specs.

Also found some utils can be merged to OpSchema so included together in
this PR

ghstack-source-id: 9ae5fb07d2c7c4fe4274bc55467d13424f655494
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/155592